requirements.txt
Flask==2.2.3
twilio==7.16.3
openai==0.26.5
python-dotenv==1.0.0